chore: migrate pnpm
