--- __tests__/sample/package.json
@@ -1,27 +1,0 @@
-{
-  "name": "sample",
-  "version": "1.0.0",
-  "private": true,
-  "scripts": {
-    "dev": "nuxt",
-    "build": "nuxt build",
-    "start": "nuxt start",
-    "generate": "nuxt generate",
-    "report:ts": "ts-node --project ../../src/ts/tsconfig.json ../../src/ts/report.ts",
-    "compare:ts": "ts-node --project ../../src/ts/tsconfig.json ../../src/ts/compare.ts",
-    "report:js": "node ../../src/js/report.js",
-    "compare:js": "node ../../src/js/compare.js"
-  },
-  "dependencies": {
-    "core-js": "^3.19.3",
-    "nuxt": "^2.15.8",
-    "vue": "^2.6.14",
-    "vue-server-renderer": "^2.6.14",
-    "vue-template-compiler": "^2.6.14",
-    "webpack": "^4.46.0"
-  },
-  "devDependencies": {},
-  "nuxtBundleAnalysis": {
-    "statsFile": ".nuxt/stats/client.json"
-  }
-}
