+++ package.json
@@ -13,10 +13,14 @@
     "actions-template"
   ],
   "scripts": {
-    "test": "vitest run",
-    "test:update": "vitest -u",
+    "test": "pnpm build && pnpm build:playground && pnpm dev:prepare && pnpm report && pnpm compare && vitest run",
+    "test:update": "pnpm build && pnpm build:playground && pnpm dev:prepare && pnpm report && pnpm compare && vitest -u",
     "minor": "npm version minor",
-    "build": "tsc --project src/tsconfig.json"
+    "build": "tsc --project src/tsconfig.json",
+    "build:playground": "pnpm --filter playground build",
+    "dev:prepare": "node scripts/test-prepare.mjs",
+    "report": "pnpm --filter playground report",
+    "compare": "pnpm --filter playground compare"
   },
   "repository": {
     "type": "git",
@@ -37,6 +41,7 @@
     "action"
   ],
   "dependencies": {
+    "@discoveryjs/json-ext": "^0.5.7",
     "filesize": "^8.0.7",
     "inquirer": "^8.2.4",
     "mkdirp": "^1.0.4"
@@ -49,7 +54,7 @@
     "esbuild-register": "^3.3.2",
     "ts-node": "^10.7.0",
     "typescript": "^4.6.2",
-    "vite": "^2.8.6",
-    "vitest": "^0.7.6"
+    "vite": "^4.2.1",
+    "vitest": "^0.29.8"
   }
 }
