+++ playground/components/NuxtLogo.vue
