+++ scripts/template/__bundle_analysis.json
