+++ vite.config.ts
@@ -1,9 +1,8 @@
-import { defineConfig } from "vitest/config";
+import { defineConfig } from 'vitest/config';
 
 export default defineConfig({
   test: {
     globals: true,
     threads: false,
-    setupFiles: ["__tests__/setup.ts"],
   },
 });
